IA-repo/requirements.txt
fastapi
uvicorn
requests
spacy
tensorflow
numpy<2
pillow
python-multipart
